bmanager3

+++ app/Http/Controllers/AdminController.php
@@ -26,7 +26,7 @@ class AdminController extends Controller
              'phone_number'=>'required ',
              'gender'=>'required',   
              'mother_name'=>'required',
-             'date_of_birth'=>'required',
+             'date_of_birth'=>'required|date_format:Y-m-d',
              'manager_address'=>'required',
             'vacations'=>'required',
             'salary'=>'required',
@@ -53,7 +53,7 @@ class AdminController extends Controller
                 $branchmanager->gender = $validator['gender'];
                 $branchmanager->mother_name = $validator['mother_name']; 
                 $branchmanager->date_of_birth = $validator['date_of_birth'];
-                 $branchmanager->address = $validator['manager_address'];
+                 $branchmanager->manager_address = $validator['manager_address'];
                  $branchmanager->vacations = $validator['vacations'];
                  $branchmanager->salary = $validator['salary'];
                  $branchmanager->rank = $validator['rank'];
@@ -83,7 +83,7 @@ class AdminController extends Controller
                       'phone_number'=>'required ',
                       'gender'=>'required',
                       'mother_name'=>'required',
-                      'date_of_birth'=>'required',
+                      'date_of_birth'=>'required|date_format:Y-m-d',
                       'manager_address'=>'required',
                      'vacations'=>'required',
                      'salary'=>'required',
@@ -104,9 +104,9 @@ class AdminController extends Controller
                 $warehousemanager->phone_number = $validator['phone_number'];
                $warehousemanager->warehouse_id = $warehouse->id;
                 $warehousemanager->gender = $validator['gender'];
-                $warehousemanager->mother_name = $validator['mother_name']; 
+                $warehousemanager->mother_name = $validator['mother_name'];
                 $warehousemanager->date_of_birth = $validator['date_of_birth'];
-                 $warehousemanager->address = $validator['manager_address'];
+                 $warehousemanager->manager_address = $validator['manager_address'];
                  $warehousemanager->vacations = $validator['vacations'];
                  $warehousemanager->salary = $validator['salary'];
                  $warehousemanager->rank = $validator['rank'];
@@ -124,8 +124,27 @@ class AdminController extends Controller
 
     public function UpdateBranch(Request $request)
     {
-
       $user = Auth::guard('admin')->user();
+      
+      $validator =Validator::make($request->all(),[
+        'branch_id'=>'numeric',
+        'address'=>'string',
+        'phone'=>'numeric|min:4',
+        'name'=>'min:5|max:255',
+        'phone_number'=> 'max:10',
+        'manager_address'=>'string',
+        'gender'=>'in:male,female',
+        'mother_name'=>'string',
+        'birth_date'=>'date_format:Y-m-d',
+        'salary'=>'string',
+        'vacations'=>'string',
+        'rank'=>'string',
+        
+    ]);
+    if ($validator->fails())
+    {
+        return response()->json($validator->errors()->toJson(),400);
+    }
       
 
       $branch = Branch::find($request->branch_id);
@@ -146,7 +165,28 @@ class AdminController extends Controller
     public function UpdateWarehouse(Request $request)
     {
 
-      
+       
+      $validator =Validator::make($request->all(),[
+        'address'=>'string',
+        'branch'=>'string',
+        'notes'=>'string',
+        'area'=>'string',
+        'phone'=>'numeric|min:4',
+        'name'=>'min:5|max:255',
+        'phone_number'=> 'max:10',
+        'manager_address'=>'string',
+        'gender'=>'in:male,female',
+        'mother_name'=>'string',
+        'birth_date'=>'date_format:Y-m-d',
+        'salary'=>'string',
+        'vacations'=>'string',
+        'rank'=>'string',
+        
+    ]);
+    if ($validator->fails())
+    {
+        return response()->json($validator->errors()->toJson(),400);
+    }
 
       $warehouse = Warehouse::find($request->warehouse_id);
       $updatedbranch= $warehouse->update($request->all());
@@ -162,6 +202,15 @@ class AdminController extends Controller
 
     public function deleteBranch(Request $request)
     {
+      $validator =Validator::make($request->all(),[
+        'branch_id'=>'required|numeric',
+    ]);
+
+    if ($validator->fails())
+    {
+        return response()->json($validator->errors()->toJson(),400);
+    }
+
         $branch = Branch::find($request->branch_id)->delete();
         $branchManager = Branch_Manager::where('branch_id', $request->branch_id)->delete();
 
@@ -171,6 +220,15 @@ class AdminController extends Controller
 
     public function deleteWarehouse(Request $request)
     {
+      $validator =Validator::make($request->all(),[
+        'warehouse_id'=>'required|numeric',
+    ]);
+
+    if ($validator->fails())
+    {
+        return response()->json($validator->errors()->toJson(),400);
+    }
+
         $Warehouse = Warehouse::find($request->warehouse_id)->delete();
       $WarehouseManager = Warehouse_Manager::where('warehouse_id', $request->warehouse_id)->delete();
